# internal/api/gen/api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for BookRequestAcceptedStatus.
const (
	Pending BookRequestAcceptedStatus = "pending"
)

// Valid indicates whether the value is a known member of the BookRequestAcceptedStatus enum.
func (e BookRequestAcceptedStatus) Valid() bool {
	switch e {
	case Pending:
		return true
	default:
		return false
	}
}

// Defines values for HealthDb.
const (
	HealthDbDown HealthDb = "down"
	HealthDbOk   HealthDb = "ok"
)

// Valid indicates whether the value is a known member of the HealthDb enum.
func (e HealthDb) Valid() bool {
	switch e {
	case HealthDbDown:
		return true
	case HealthDbOk:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatus.
const (
	HealthStatusOk HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case HealthStatusOk:
		return true
	default:
		return false
	}
}

// BookRequest defines model for BookRequest.
type BookRequest struct {
	GutenbergId int     `json:"gutenbergId"`
	Language    *string `json:"language,omitempty"`

	// Title 관리자가 카탈로그에서 확인한 제목. 수집 결과 검증에 쓴다
	Title string `json:"title"`
}

// BookRequestAccepted defines model for BookRequestAccepted.
type BookRequestAccepted struct {
	BookId int64                     `json:"bookId"`
	Status BookRequestAcceptedStatus `json:"status"`
}

// BookRequestAcceptedStatus defines model for BookRequestAccepted.Status.
type BookRequestAcceptedStatus string

// Chapter defines model for Chapter.
type Chapter struct {
	Idx int `json:"idx"`

	// Title 비어 있을 수 있다. 파서가 제목을 못 찾거나 후처리로 비운 경우다
	Title string `json:"title"`
}

// ChapterView defines model for ChapterView.
type ChapterView struct {
	Chapter    Chapter     `json:"chapter"`
	Paragraphs []Paragraph `json:"paragraphs"`

	// TotalChapters 이 revision의 전체 챕터 수. 이전·다음 링크 판단에 쓴다
	TotalChapters int `json:"totalChapters"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// Health defines model for Health.
type Health struct {
	// Db Postgres Ping 결과
	Db HealthDb `json:"db"`

	// Status 프로세스가 살아 있으면 항상 ok
	Status HealthStatus `json:"status"`

	// Version 빌드 버전. 개발 중에는 dev
	Version string `json:"version"`
}

// HealthDb Postgres Ping 결과
type HealthDb string

// HealthStatus 프로세스가 살아 있으면 항상 ok
type HealthStatus string

// Paragraph defines model for Paragraph.
type Paragraph struct {
	SourceText string `json:"sourceText"`

	// StableId 번역이 붙는 키 (ADR-004/016). paragraphs.id가 아니다
	StableId string `json:"stableId"`
}

// RequestBookJSONRequestBody defines body for RequestBook for application/json ContentType.
type RequestBookJSONRequestBody = BookRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// RequestBook 도서 수집 지시
	// (POST /admin/books)
	RequestBook(w http.ResponseWriter, r *http.Request)
	// GetBookChapter 활성 revision의 챕터 하나와 그 문단들
	// (GET /books/{gutenbergId}/chapters/{idx})
	GetBookChapter(w http.ResponseWriter, r *http.Request, gutenbergId int, idx int)
	// GetHealthz 서비스와 DB 연결 상태
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// RequestBook 도서 수집 지시
// (POST /admin/books)
func (_ Unimplemented) RequestBook(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetBookChapter 활성 revision의 챕터 하나와 그 문단들
// (GET /books/{gutenbergId}/chapters/{idx})
func (_ Unimplemented) GetBookChapter(w http.ResponseWriter, r *http.Request, gutenbergId int, idx int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetHealthz 서비스와 DB 연결 상태
// (GET /healthz)
func (_ Unimplemented) GetHealthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// RequestBook operation middleware
func (siw *ServerInterfaceWrapper) RequestBook(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RequestBook(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBookChapter operation middleware
func (siw *ServerInterfaceWrapper) GetBookChapter(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "gutenbergId" -------------
	var gutenbergId int

	err = runtime.BindStyledParameterWithOptions("simple", "gutenbergId", chi.URLParam(r, "gutenbergId"), &gutenbergId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "gutenbergId", Err: err})
		return
	}

	// ------------- Path parameter "idx" -------------
	var idx int

	err = runtime.BindStyledParameterWithOptions("simple", "idx", chi.URLParam(r, "idx"), &idx, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "idx", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBookChapter(w, r, gutenbergId, idx)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealthz)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/books/{gutenbergId}/chapters/{idx}", wrapper.GetBookChapter)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/books", wrapper.RequestBook)
	})

	return r
}

type RequestBookRequestObject struct {
	Body *RequestBookJSONRequestBody
}

type RequestBookResponseObject interface {
	VisitRequestBookResponse(w http.ResponseWriter) error
}

type RequestBook202JSONResponse BookRequestAccepted

func (response RequestBook202JSONResponse) VisitRequestBookResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(202)
	_, err := buf.WriteTo(w)
	return err
}

type RequestBook401JSONResponse Error

func (response RequestBook401JSONResponse) VisitRequestBookResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type RequestBook409JSONResponse Error

func (response RequestBook409JSONResponse) VisitRequestBookResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type GetBookChapterRequestObject struct {
	GutenbergId int `json:"gutenbergId"`
	Idx         int `json:"idx"`
}

type GetBookChapterResponseObject interface {
	VisitGetBookChapterResponse(w http.ResponseWriter) error
}

type GetBookChapter200JSONResponse ChapterView

func (response GetBookChapter200JSONResponse) VisitGetBookChapterResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetBookChapter404JSONResponse Error

func (response GetBookChapter404JSONResponse) VisitGetBookChapterResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse Health

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// RequestBook 도서 수집 지시
	// (POST /admin/books)
	RequestBook(ctx context.Context, request RequestBookRequestObject) (RequestBookResponseObject, error)
	// GetBookChapter 활성 revision의 챕터 하나와 그 문단들
	// (GET /books/{gutenbergId}/chapters/{idx})
	GetBookChapter(ctx context.Context, request GetBookChapterRequestObject) (GetBookChapterResponseObject, error)
	// GetHealthz 서비스와 DB 연결 상태
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// RequestBook operation middleware
func (sh *strictHandler) RequestBook(w http.ResponseWriter, r *http.Request) {
	var request RequestBookRequestObject

	var body RequestBookJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RequestBook(ctx, request.(RequestBookRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RequestBook")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RequestBookResponseObject); ok {
		if err := validResponse.VisitRequestBookResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetBookChapter operation middleware
func (sh *strictHandler) GetBookChapter(w http.ResponseWriter, r *http.Request, gutenbergId int, idx int) {
	var request GetBookChapterRequestObject

	request.GutenbergId = gutenbergId
	request.Idx = idx

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetBookChapter(ctx, request.(GetBookChapterRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetBookChapter")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetBookChapterResponseObject); ok {
		if err := validResponse.VisitGetBookChapterResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(w http.ResponseWriter, r *http.Request) {
	var request GetHealthzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}