# openapi.yaml
# API 계약의 단일 원본 (ADR-009).
# 여기부터 고치고 `make generate`를 돌린다. Go 핸들러나 TS 클라이언트를 먼저 고치면 덮어써진다.
openapi: 3.0.3
info:
  title: reclassic API
  version: 0.1.0
  description: |
    Project Gutenberg 도서를 장·문단으로 분해하고 문단별 번역을 검수·공개하는 서비스의 API.
    지금은 골격 슬라이스라 헬스체크 하나뿐이다.

servers:
  - url: http://localhost:8080
    description: 로컬 개발

paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: 서비스와 DB 연결 상태
      description: db 필드는 실제 커넥션 Ping 결과다. 상수가 아니다.
      responses:
        "200":
          description: 서비스가 응답한다. db가 down이어도 200이다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  # 잠정 계약이다. 최종은 번역 프로젝트 기반의 GET /projects/{id}/chapters/{idx}이고
  # (ARCHITECTURE "API 계약"), 그건 translation_projects가 생겨야 성립한다.
  # 번역 슬라이스에서 교체하며, 그때 이 엔드포인트를 남길지 없앨지 함께 정한다.
  #
  # 번역 관련 필드는 넣지 않는다. approvedTranslation을 항상 null로 채우면
  # 계약이 거짓말을 하게 된다. 없는 것은 스키마에 없어야 한다.
  /books/{gutenbergId}/chapters/{idx}:
    get:
      operationId: getBookChapter
      summary: 활성 revision의 챕터 하나와 그 문단들
      description: |
        활성 revision이 없으면(수집 실패, 게이트에 걸려 needs_review인 경우) 404다.
        읽기 화면은 챕터 단위로 자른다 — 책 하나에 문단이 5,000개를 넘길 수 있다.
      parameters:
        - name: gutenbergId
          in: path
          required: true
          description: Project Gutenberg 도서 번호
          schema:
            type: integer
            minimum: 1
        - name: idx
          in: path
          required: true
          description: 0부터 시작하는 챕터 순번
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: 챕터와 문단
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterView"
        "404":
          description: 활성 revision이 없거나 해당 챕터가 없다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/books:
    post:
      operationId: requestBook
      summary: 도서 수집 지시
      description: |
        books 행 생성과 FetchSource 잡 등록을 **한 트랜잭션 안에서** 한다.
        "책은 생성됐는데 잡 등록이 실패"하는 틈이 없다 — ADR-003이 River를 고른 근거다.

        수집은 비동기라 202다. 결과는 books.status로 확인한다.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
      responses:
        "202":
          description: 수집을 지시했다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookRequestAccepted"
        "401":
          description: 관리자 토큰이 없거나 틀렸다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 이미 수집이 지시된 책이다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    # 임시 운영 가드다. 세션 인증은 아직 열린 질문이라(AGENTS.md)
    # 그때까지 관리자 엔드포인트를 무인증으로 열어 두지 않기 위한 것뿐이다.
    # 사용자 인증(제안자·검수자)은 이 방식으로 하지 않는다.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token

  schemas:
    Health:
      type: object
      required: [status, db, version]
      properties:
        status:
          type: string
          enum: [ok]
          description: 프로세스가 살아 있으면 항상 ok
        db:
          type: string
          enum: [ok, down]
          description: Postgres Ping 결과
        version:
          type: string
          description: 빌드 버전. 개발 중에는 dev

    ChapterView:
      type: object
      required: [chapter, paragraphs, totalChapters]
      properties:
        chapter:
          $ref: "#/components/schemas/Chapter"
        paragraphs:
          type: array
          items:
            $ref: "#/components/schemas/Paragraph"
        totalChapters:
          type: integer
          description: 이 revision의 전체 챕터 수. 이전·다음 링크 판단에 쓴다

    Chapter:
      type: object
      required: [idx, title]
      properties:
        idx:
          type: integer
        title:
          type: string
          description: 비어 있을 수 있다. 파서가 제목을 못 찾거나 후처리로 비운 경우다

    Paragraph:
      type: object
      required: [stableId, sourceText]
      properties:
        stableId:
          type: string
          description: 번역이 붙는 키 (ADR-004/016). paragraphs.id가 아니다
        sourceText:
          type: string

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string

    BookRequest:
      type: object
      required: [gutenbergId, title]
      properties:
        gutenbergId:
          type: integer
          minimum: 1
        title:
          type: string
          minLength: 1
          description: 관리자가 카탈로그에서 확인한 제목. 수집 결과 검증에 쓴다
        language:
          type: string
          default: en

    BookRequestAccepted:
      type: object
      required: [bookId, status]
      properties:
        bookId:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending]